Move Approval API into MergeRequestApprovals
Breaking change, see https://github.com/jdalrymple/gitbeaker/pull/1233#issuecomment-709262998
