+++ packages/gitbeaker-core/src/services/MergeRequests.ts
@@ -15,12 +15,6 @@ export interface AcceptMergeRequestOptions {
   sha?: string;
 }
 
-export type ApprovalRulesRequestOptions = {
-  userIds?: number[];
-  groupIds?: number[];
-  protectedBranchIds?: number[];
-};
-
 export interface ShowMergeRequestOptions {
   renderHtml?: boolean;
   includeDivergedCommitsCount?: true;
@@ -89,29 +83,6 @@ export class MergeRequests extends BaseService {
     return RequestHelper.put(this, `projects/${pId}/merge_requests/${mIid}/merge`, options);
   }
 
-  addApprovalRule(
-    projectId: string | number,
-    name: string,
-    approvalsRequired: number,
-    {
-      mergerequestIid,
-      ...options
-    }: { mergerequestIid?: number } & ApprovalRulesRequestOptions & BaseRequestOptions,
-  ) {
-    const pId = encodeURIComponent(projectId);
-
-    let url;
-
-    if (mergerequestIid) {
-      const mIid = encodeURIComponent(mergerequestIid);
-      url = `projects/${pId}/merge_requests/${mIid}/approval_rules`;
-    } else {
-      url = `projects/${pId}/approval_rules`;
-    }
-
-    return RequestHelper.post(this, url, { name, approvalsRequired, ...options });
-  }
-
   addSpentTime(
     projectId: string | number,
     mergerequestIid: number,
@@ -158,85 +129,6 @@ export class MergeRequests extends BaseService {
     }
 
     return RequestHelper.get(this, url, options);
-  }
-
-  approve(
-    projectId: string | number,
-    mergerequestIid: number,
-    options?: { sha?: string } & BaseRequestOptions,
-  ) {
-    const [pId, mIid] = [projectId, mergerequestIid].map(encodeURIComponent);
-
-    return RequestHelper.post(this, `projects/${pId}/merge_requests/${mIid}/approve`, options);
-  }
-
-  approvals(
-    projectId: string | number,
-    { mergerequestIid, ...options }: { mergerequestIid?: number } & BaseRequestOptions = {},
-  ) {
-    const pId = encodeURIComponent(projectId);
-
-    let url;
-
-    if (mergerequestIid) {
-      const mIid = encodeURIComponent(mergerequestIid);
-      url = `projects/${pId}/merge_requests/${mIid}/approvals`;
-    } else {
-      url = `projects/${pId}/approvals`;
-    }
-
-    return RequestHelper.get(this, url, options);
-  }
-
-  approvalRules(
-    projectId: string | number,
-    { mergerequestIid, ...options }: { mergerequestIid?: number } & BaseRequestOptions = {},
-  ) {
-    const pId = encodeURIComponent(projectId);
-
-    let url;
-
-    if (mergerequestIid) {
-      const mIid = encodeURIComponent(mergerequestIid);
-      url = `projects/${pId}/merge_requests/${mIid}/approval_rules`;
-    } else {
-      url = `projects/${pId}/approval_rules`;
-    }
-    return RequestHelper.get(this, url, options);
-  }
-
-  approvalState(
-    projectId: string | number,
-    mergerequestIid: number,
-    options?: { sha?: string } & BaseRequestOptions,
-  ) {
-    const [pId, mIid] = [projectId, mergerequestIid].map(encodeURIComponent);
-
-    return RequestHelper.get(
-      this,
-      `projects/${pId}/merge_requests/${mIid}/approval_state`,
-      options,
-    );
-  }
-
-  approvers(
-    projectId: string | number,
-    approverIds: number[],
-    approverGroupIds: (string | number)[],
-    { mergerequestIid, ...options }: { mergerequestIid?: number } & BaseRequestOptions = {},
-  ) {
-    const pId = encodeURIComponent(projectId);
-
-    let url;
-
-    if (mergerequestIid) {
-      const mIid = encodeURIComponent(mergerequestIid);
-      url = `projects/${pId}/merge_requests/${mIid}/approvers`;
-    } else {
-      url = `projects/${pId}/approvers`;
-    }
-
-    return RequestHelper.put(this, url, { approverIds, approverGroupIds, ...options });
   }
 
   cancelOnPipelineSucess(projectId: string | number, mergerequestIid: number, options?: Sudo) {
@@ -295,48 +187,6 @@ export class MergeRequests extends BaseService {
     return RequestHelper.put(this, `projects/${pId}/merge_requests/${mIid}`, options);
   }
 
-  editApprovals(
-    projectId: string | number,
-    { mergerequestIid, ...options }: { mergerequestIid?: number } & BaseRequestOptions = {},
-  ) {
-    const pId = encodeURIComponent(projectId);
-
-    let url;
-
-    if (mergerequestIid) {
-      const mIid = encodeURIComponent(mergerequestIid);
-      url = `projects/${pId}/merge_requests/${mIid}/approvals`;
-    } else {
-      url = `projects/${pId}/approvals`;
-    }
-
-    return RequestHelper.post(this, url, options);
-  }
-
-  editApprovalRule(
-    projectId: string | number,
-    approvalRuleId: number,
-    name: string,
-    approvalsRequired: number,
-    {
-      mergerequestIid,
-      ...options
-    }: { mergerequestIid?: number } & ApprovalRulesRequestOptions & BaseRequestOptions = {},
-  ) {
-    const [pId, aId] = [projectId, approvalRuleId].map(encodeURIComponent);
-
-    let url;
-
-    if (mergerequestIid) {
-      const mIid = encodeURIComponent(mergerequestIid);
-      url = `projects/${pId}/merge_requests/${mIid}/approval_rules/${aId}`;
-    } else {
-      url = `projects/${pId}/approval_rules/${aId}`;
-    }
-
-    return RequestHelper.put(this, url, { name, approvalsRequired, ...options });
-  }
-
   participants(projectId: string | number, mergerequestIid: number, options?: Sudo) {
     const [pId, mIid] = [projectId, mergerequestIid].map(encodeURIComponent);
 
@@ -359,25 +209,6 @@ export class MergeRequests extends BaseService {
     const [pId, mIid] = [projectId, mergerequestIid].map(encodeURIComponent);
 
     return RequestHelper.del(this, `projects/${pId}/merge_requests/${mIid}`, options);
-  }
-
-  removeApprovalRule(
-    projectId: string | number,
-    approvalRuleId: number,
-    { mergerequestIid, ...options }: { mergerequestIid?: number } & BaseRequestOptions = {},
-  ) {
-    const [pId, aId] = [projectId, approvalRuleId].map(encodeURIComponent);
-
-    let url;
-
-    if (mergerequestIid) {
-      const mIid = encodeURIComponent(mergerequestIid);
-      url = `projects/${pId}/merge_requests/${mIid}/approval_rules/${aId}`;
-    } else {
-      url = `projects/${pId}/approval_rules/${aId}`;
-    }
-
-    return RequestHelper.del(this, url, options);
   }
 
   resetSpentTime(projectId: string | number, mergerequestIid: number, options?: Sudo) {
@@ -436,12 +267,6 @@ export class MergeRequests extends BaseService {
     const [pId, mIid] = [projectId, mergerequestIid].map(encodeURIComponent);
 
     return RequestHelper.get(this, `projects/${pId}/merge_requests/${mIid}/versions`, options);
-  }
-
-  unapprove(projectId: string | number, mergerequestIid: number, options?: Sudo) {
-    const [pId, mIid] = [projectId, mergerequestIid].map(encodeURIComponent);
-
-    return RequestHelper.post(this, `projects/${pId}/merge_requests/${mIid}/unapprove`, options);
   }
 
   unsubscribe(projectId: string | number, mergerequestIid: number, options?: Sudo) {
